// packages/web/src/shared/stores/demo/types.ts
export interface PendingScorerData {
  __identity: string
  displayName: string
  birthday?: string
}